inject: add loose regex
