+++ test/fixtures/expected-regex.md
@@ -1,0 +1,5 @@
+<!-- START main-header -->
+# best header ever
+<!-- END main-header -->
+
+## another section
