+++ test/fixtures/file-regex.md
@@ -1,0 +1,4 @@
+<!-- START main-header -->
+<!-- END main-header -->
+
+## another section
